requirements.txt
requests==2.32.3
PyJWT==2.10.1
mitmproxy==11.1.3
